image_compression/decoder/requirements.txt
bitstring
numpy
opencv-python
Pillow
scipy
